core: clean up warnings

+++ src/audio_core/dsp_interface.cpp
@@ -45,7 +45,7 @@ void DspInterface::OutputFrame(StereoFrame16& frame) {
         return;
 
     // Implementation of the hardware volume slider with a dynamic range of 60 dB
-    float volume_scale_factor = std::exp(6.90775 * Settings::values.volume) * 0.001;
+    double volume_scale_factor = std::exp(6.90775 * Settings::values.volume) * 0.001;
     for (size_t i = 0; i < frame.size(); i++) {
         frame[i][0] = static_cast<s16>(frame[i][0] * volume_scale_factor);
         frame[i][1] = static_cast<s16>(frame[i][1] * volume_scale_factor);
